move TensorMeta implementation from tensor_impl.cpp to tensor_meta.cpp
