+++ oneflow/core/framework/tensor_impl.cpp
@@ -24,7 +24,6 @@ limitations under the License.
 #include "oneflow/core/job/parallel_desc.h"
 #include "oneflow/core/job/sbp_parallel.h"
 #include "oneflow/core/functional/functional.h"
-#include "oneflow/core/framework/device.h"
 #include "oneflow/core/framework/dtype.h"
 #include "oneflow/core/eager/eager_blob_object.h"
 #include "oneflow/core/eager/local_dep_object.h"
@@ -159,53 +158,6 @@ Maybe<void> EagerMirroredTensorImpl::RegisterStorageDeleteHook(const std::functi
   CHECK_OR_RETURN(eager_blob_object_) << "EagerBlobObject has not initialized";
   eager_blob_object_->RegisterStorageDeleteHook(hook);
   return Maybe<void>::Ok();
-}
-
-MirroredTensorMeta::MirroredTensorMeta()
-    : TensorMeta(std::make_shared<const Shape>(), DataType::kInvalidDataType),
-      device_(Symbol<Device>()),
-      stride_(std::make_shared<const Stride>()),
-      is_contiguous_(false),
-      storage_offset_(0) {}
-
-MirroredTensorMeta::MirroredTensorMeta(const std::shared_ptr<const Shape>& shape, DataType dtype,
-                                       Symbol<Device> device)
-    : TensorMeta(shape, dtype),
-      device_(device),
-      stride_(std::make_shared<const Stride>(*shape)),
-      is_contiguous_(false),
-      storage_offset_(0) {}
-
-MirroredTensorMeta::MirroredTensorMeta(const std::shared_ptr<const Shape>& shape, DataType dtype,
-                                       Symbol<Device> device,
-                                       const std::shared_ptr<const Stride>& stride,
-                                       int64_t storage_offset)
-    : TensorMeta(shape, dtype), device_(device), stride_(stride), storage_offset_(storage_offset) {
-  is_contiguous_ = IsContiguous(*shape, *stride);
-}
-
-bool MirroredTensorMeta::operator==(const MirroredTensorMeta& other) const {
-  // It's correct to ignore is_dynamic_ field.
-  return *this->shape_ptr() == *other.shape_ptr() && this->dtype() == other.dtype()
-         && *this->device() == *other.device() && this->stride() == other.stride();
-}
-
-size_t MirroredTensorMeta::CalcHashValue() const {
-  // It's correct to ignore is_dynamic_ field.
-  return std::hash<Shape>()(*shape_ptr()) ^ std::hash<DataType>()(dtype())
-         ^ std::hash<Device>()(*device()) ^ std::hash<Stride>()(stride());
-}
-
-bool ConsistentTensorMeta::operator==(const ConsistentTensorMeta& other) const {
-  // It's correct to ignore is_dynamic_ field.
-  return *this->shape_ptr() == *other.shape_ptr() && this->dtype() == other.dtype()
-         && this->nd_sbp() == other.nd_sbp() && this->parallel_desc() == other.parallel_desc();
-}
-
-size_t ConsistentTensorMeta::CalcHashValue() const {
-  return std::hash<Shape>()(*shape_ptr()) ^ std::hash<DataType>()(dtype())
-         ^ std::hash<Symbol<NdSbp>>()(nd_sbp())
-         ^ std::hash<Symbol<ParallelDesc>>()(parallel_desc());
 }
 
 Maybe<ConsistentTensorImpl> LazyConsistentTensorImpl::detach() const {
